+++ scripts/macros/2014/classFeatures/illrigger/interdictBoons/hellsight.js
@@ -1,0 +1,26 @@
+import {itemUtils} from '../../../../../utils.js';
+async function added({trigger: {entity: item}}) {
+    await itemUtils.correctActivityItemConsumption(item, ['use'], 'balefulInterdict');
+}
+export let interdictBoonHellsight = {
+    name: 'Interdict Boons: Hellsight',
+    version: '1.7.71',
+    rules: 'legacy',
+    item: [
+        {
+            pass: 'created',
+            macro: added,
+            priority: 50
+        },
+        {
+            pass: 'itemMedkit',
+            macro: added,
+            priority: 50
+        },
+        {
+            pass: 'actorMunch',
+            macro: added,
+            priority: 50
+        }
+    ]
+};