+++ scripts/macros/2014/monsterFeatures/giant/fireGiantOfEvilFire/boltOfImix.js
@@ -1,0 +1,8 @@
+export let boltOfImix = {
+    name: 'Bolt of Imix',
+    version: '1.3.72',
+    rules: 'legacy',
+    monster: {
+        name: 'Fire Giant of Evil Fire'
+    }
+};