+++ scripts/macros/2014/monsterFeatures/giant/fireGiantOfEvilFire/searingScepter.js
@@ -1,0 +1,8 @@
+export let searingScepter = {
+    name: 'Searing Scepter',
+    version: '1.3.72',
+    rules: 'legacy',
+    monster: {
+        name: 'Fire Giant of Evil Fire'
+    }
+};