+++ scripts/macros/2014/monsterFeatures/giant/fireGiantOfEvilFire/shrapnelExplosion.js
@@ -1,0 +1,8 @@
+export let shrapnelExplosion = {
+    name: 'Shrapnel Explosion',
+    version: '1.3.72',
+    rules: 'legacy',
+    monster: {
+        name: 'Fire Giant of Evil Fire'
+    }
+};